Calc total bill costs
Basically using user input this will calculate the total costs of all the bills for the user, find the month with the highest amount of bills paid, and find the month with the lowest amount of bills paid. This will be implemented in the "activity" tab of our project where all the information will be shown for users to track their finances.